break up calculators into their own files
